Reduced the lower-bound alert threshold temperatures
for hard drives and solid-state drives.

Previously, the thresholds were 20C, 17C, 14C, 8C.  However, we have
been routinely seeing 20C on SSD in the datacenter, which is a normal
operating temperature.

Given that the lower-bound operating temperature of both SSD and HDD is
around 0C to 5C, we have adjusted the alert thresholds to 10C, 8C, 6C,
5C.